les-prono/admin_scripts/requirements.txt
alabaster==0.7.12
appdirs==1.4.4
attrs==19.3.0
Babel==2.8.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
docutils==0.16
idna==2.10
imagesize==1.2.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
numpy==1.18.4
opencv-python-headless==4.2.0.34
packaging==20.4
pandas==1.0.3
pathspec==0.8.0
pockets==0.9.1
pvlib==0.7.2
Pygments==2.6.1
pyparsing==2.4.7
python-crontab==2.4.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.24.0
rstcheck==3.3.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.14.0
sklearn-contrib-py-earth==0.1.0
snowballstemmer==2.0.0
Sphinx==3.1.2
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
threadpoolctl==2.1.0
timezonefinder==4.3.1
toml==0.10.1
tqdm==4.46.1
typed-ast==1.4.1
urllib3==1.25.9
